Added roles checking (#7)
* feat(common): added `CurrentUser` decorator

* feat(user): user can only be deleted by admin or user himself

* feat(user): user can only be updated by himself

* feat(tutor): profile can only be updated by owner of this profile

* feat(tutor): some endpoints get `userId` from access token

It secures endpoints (only owner of profile can interact with it)

* feat(tutor): updated documentation

* feat(student): some endpoints get `userId` from access token

* feat(student): updated documentation

* feat(user): some endpoints get `userId` from access token

* feat(roles): added `RolesGuard` and `Roles` decorator

* feat(tutor): some endpoints are protected with `Roles` decorator

* feat(student): some endpoints are protected with `Roles` decorator

+++ src/user/user.service.ts
@@ -1,9 +1,10 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '@prisma/prisma.service';
-import { User } from '@prisma/client';
+import { Role, User } from '@prisma/client';
 import { genSalt, hash } from 'bcrypt';
 import { CreateUserDTO, UpdateUserDTO } from '@user/dto';
 import { DeleteResponse } from '@user/responses';
+import { JwtPayload } from '@auth/interfaces';
 
 @Injectable()
 export class UserService {
@@ -23,15 +24,12 @@ export class UserService {
   }
 
   async update(id: string, user: UpdateUserDTO): Promise<User> {
-    await this.findOne(id, true); // Trying to find a user to update
+    // Trying to find a user to update
+    await this.findOne(id, true);
 
     return this.prisma.user.update({
-      where: {
-        id: id,
-      },
-      data: {
-        ...user,
-      },
+      where: { id: id },
+      data: { ...user },
     });
   }
 
@@ -53,15 +51,14 @@ export class UserService {
     return this.prisma.user.findMany();
   }
 
-  async delete(id: string): Promise<DeleteResponse> {
-    //   TODO: add check if user tries to delete himself of it is admin
+  async delete(id: string, currentUser: JwtPayload): Promise<DeleteResponse> {
+    if (currentUser.id !== id || currentUser.role !== Role.ADMIN) {
+      throw new ForbiddenException("User can't be deleted");
+    }
+
     return this.prisma.user.delete({
-      where: {
-        id: id,
-      },
-      select: {
-        id: true,
-      },
+      where: { id: id },
+      select: { id: true },
     });
   }
